Interim Commit - 4 Clean up test domain initialization
