+++ tests/adapters/model/sqlalchemy_model/sqlite/conftest.py
@@ -1,27 +1,11 @@
-import os
-
 import pytest
 
-
-def initialize_domain():
-    from protean.domain import Domain
-
-    domain = Domain(__file__, "SQLAlchemy Test - SQLite")
-
-    # Construct relative path to config file
-    current_path = os.path.abspath(os.path.dirname(__file__))
-    config_path = os.path.join(current_path, "./config.py")
-
-    if os.path.exists(config_path):
-        domain.config.from_pyfile(config_path)
-
-    return domain
+from tests.shared import initialize_domain
 
 
 @pytest.fixture(autouse=True)
 def test_domain():
-    domain = initialize_domain()
-    domain.reinitialize()
+    domain = initialize_domain(__file__)
 
     with domain.domain_context():
         yield domain
@@ -29,6 +13,7 @@ def test_domain():
 
 @pytest.fixture(scope="session", autouse=True)
 def setup_db():
+    domain = initialize_domain(__file__)
     with domain.domain_context():
         # Create all associated tables
         from .elements import ComplexUser, Person, Provider, ProviderCustomModel, User
