+++ tests/conftest.py
@@ -4,10 +4,7 @@
 """
 
 import os
-
 import pytest
-
-from uuid import UUID
 
 
 def pytest_addoption(parser):
@@ -197,7 +194,9 @@ def test_domain(db_config, store_config, request):
         domain.config["DATABASES"]["default"] = db_config
         domain.config["EVENT_STORE"] = store_config
 
+        # Always reinitialize the domain after config changes
         domain.reinitialize()
+
         with domain.domain_context():
             yield domain
 
@@ -244,13 +243,3 @@ def run_around_tests(test_domain):
             cache.flush_all()
 
         test_domain.event_store.store._data_reset()
-
-
-def assert_str_is_uuid(value: str) -> None:
-    try:
-        UUID(value)
-    except ValueError:
-        pytest.fail("Invalid UUID")
-
-
-pytest.assert_str_is_uuid = assert_str_is_uuid
